# core/clist_sort: rewrite