--- a/tests/bench/runtime_coreapis/tests/01-run.py
+++ b/tests/bench/runtime_coreapis/tests/01-run.py
@@ -27,6 +27,9 @@ def testfunc(child):
     child.expect(BENCHMARK_REGEXP.format(func="thread flags set/wait one"), timeout=TIMEOUT)
     child.expect(BENCHMARK_REGEXP.format(func=r"msg_try_receive\(\)"), timeout=TIMEOUT)
     child.expect(BENCHMARK_REGEXP.format(func=r"msg_avail\(\)"))
+    child.expect(BENCHMARK_REGEXP.format(func=r"clist_sort\(\) \(\d+ nodes\)", timeout=TIMEOUT))
+    child.expect(BENCHMARK_REGEXP.format(func=r"clist_sort\(\) \(\d+ nodes\)", timeout=TIMEOUT))
+    child.expect(BENCHMARK_REGEXP.format(func=r"clist_sort\(\) \(\d+ nodes\)", timeout=TIMEOUT))
     child.expect_exact('[SUCCESS]')