### Contribution description

This rewrites `clist_sort()` with a different implementation of a merge sort. The goal is that the code is more readable and passes `-fanalyzer` without GCC barking, but it turns out to be a bit faster as well.

### Testing procedure

This currently includes two other PRs for better testing, namely:

- [ ] https://github.com/RIOT-OS/RIOT/pull/21403
- [ ] https://github.com/RIOT-OS/RIOT/pull/21402

(This PR needs to wait for those to be merged first and rebased then.)

### Issues/PRs references

More complex but faster alternative to https://github.com/RIOT-OS/RIOT/pull/21398